+++ integration_tests/models/dbt_utils_overrides/generic_tests/data_unique_combination_of_columns_quoted.sql
@@ -1,0 +1,3 @@
+-- data_unique_combination_of_columns is quoted (MONTH is reserved). The tests in schema.yml use quoted column names.
+
+select * from {{ ref('data_unique_combination_of_columns') }}